[dataset] recover pending dataset delay upon reboot as router/child
This commit updates the code to allow recovery of pending dataset
delay timer when a device reboots and attempts to restore its
previous role as a router or child.

Upon reboot, a device can restore a saved pending dataset from
non-volatile settings but it cannot resume the delay timer since it
does not know when the dataset was originally saved.

This commit changes the code such that if the pending dataset delay
timer is not running, the "Pending Timestamp TLV" is not appended in
MLE messages, specifically in "Data Request" message. This ensures
that other devices include the full pending dataset (if any) in their
response, allowing the rebooted device to learn the current delay
timer.

This commit also includes the addition of
`test-033-pending-dataset-after-reset.py` to validate this behavior.