[test] skip packet verification on failure
The packet verification is only skipped when the test case was skipped,
and when the test already failed, it still attempts to perform the
verification. This commit marks the verification should be skipped when
the test either failed or was skipped, so that it won't attempt to
do packet verification in both cases. This can reduce some confusion
when inspecting test failures.