[mle] simplify graceful detach mechanism and reuse `mAttachTimer`
This commit updates the graceful detach mechanism in `Mle`. A boolean
`mDetachingGracefully` is added to track when the device is
performing a "graceful detach". The `mAttachTimer` is reused for
scheduling the detach delay, as any attach operation can be stopped
during detach, and the timer can be reused. This simplifies the code
and removes the need to use a specific timer for the graceful detach
process.