[udp] track netif ID in socket and simplify handling of backbone sockets
This commit updates `otUdpSocket` to track the associated network
interface identifier to which the socket is bound. This simplifies
the code for tracking backbone sockets, particularly how
`Udp::HandlePayload()` finds a matching socket.

Previously, backbone sockets were placed at the end of the `mSockets`
linked-list. The code tracked and updated a pointer to the entry
before the first backbone socket in the list. This commit removes the
need for this additional tracking and simplifies the implementation.