[trel] add otSys API to init/deinit TREL
The current TREL implementation uses TREL URL to pass in the TREL
interface at system start. This works for devices that have a fixed
interface for TREL.

There are devices that can dynamically switch infrastructure link, for
those devices it's possible that infra link changes from ethernet to
wifi and TREL should be re-configured on the new interface.