[mdns] simplify response and query preparation by passing `Context`
This commit simplifies the process of preparing response and query
messages in the mDNS module. Instead of passing `TxMessage` and
`aNow` separately,  a reference to `EntryContext` or `CacheContext`
is now passed directly to `PrepareResponse()` and other related
methods.

This direct approach streamlines the code and facilitates the
potential inclusion of additional fields in `EntryContext` in the
future.