[mac] correct the version 2015 frame format generation
If a 15.4 frame with version 2015 is generated using the rule of
No. 12 ~ 14 (no source PanId presents) of the method `Frame::PrepareHeadersIn()`,
a zero-valued source PanId will be allocated for the frame. When
the frame parser parses the frame, it gets a zero-valued source
PanId from the frame. The set source PanId is inconsistent with
the parsed source PanId.

This commit corrects this inconsistency and updates the unit test.