[child-supervision] Support Short Supervision Interval TLV
A Wake-up End Device uses the child supervision mechanism
to assure a proper time synchronization of the enhanced CSL
link, and to quickly react on the link loss in order to
return to wake-up frame sniffing, and be able to recover
the connection with no delay.

To achieve the above goals, the WED needs to use the child
supervision interval shorter than 1s that is currently the
minimum value possible to encode in Supervision Interval TLV.
To overcome this limitation, the Short Supervision Interval
TLV element is being introduced, which has the same meaning
as Supervision Interval TLV, except that it carries the
interval in units of 100ms, instead of seconds.

Implement handling Short Supervision Interval TLV in Thread
routers. Subsequent commits will add support for the short
supervision interval on WEDs being attached to Wake-up
Parents.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>
Co-authored-by: Eduardo Montoya <eduardo.montoya@nordicsemi.no>