[secure-transport] fix `Send()` method to send entire message
This commit simplifies and fixes the `Send()` method by sending the
entire `aMessage` content and removing the extra `aLength` input
parameter. In all calls, the input `aLength` was set to the message's
length. The previous implementation of `Send()` did not correctly
handle cases where the passed-in length was longer than the available
bytes in the given message.

This commit also updates the documentation of `Send()` to clarify the
behavior regarding the transfer of ownership of `aMessage`.

Additionally, this commit simplifies `CoapSecure` and `BleSecure`
where `Send()` is called.