[mle] do not request link if not recognized (#11007)
If the new router is not in peer's route TLV, the link request would
be ignored, causing the link establishment stuck at the Link Request
stage until timeout when a new attempt could proceed. This commit adds
a check on the initiator to make sure the peer router's Route TLV
contains the initiator's router id before sending the link request,
otherwise it refrains from initiating the link establishment.