[mle] abort delayed Link Request after promotion
A node clears its RouterTable after promoting to a router. A Link
Request scheduled when it was a child is still sent out. If the other
router happen to be able to process the request, it will respond with
Link Accept and Request, which will be dropped because the newly
promoted router doesn't think it came frame a router in the LinkRequest
state.

This commit aborts the delayed Link Request after promotion. It should
not sent out the Link Request and update the router state to LinkRequest
because it is not sure whether the other router reconizes it yet.