[thread-netif] do not close `Tmf::SecureAgent` on `Down()`
This commit updates the code so that `Tmf::SecureAgent` is not closed
when the Thread network interface is brought down. The `SecureAgent`
is directly controlled and used by either the `BorderAgent` or
`Commissioner` during network operation, which should continue to
function even if the Thread network interface is offline. It may also
be used by the `Joiner`, but in that case, it will be closed by the
`Joiner` itself upon finishing the join attempt.