[secure-transport] simplify how `Connect/Receive` callbacks are set (#11050)
This commit updates and simplifies how `Connect` and `Receive`
callbacks are set on `SecureTransport` and `CoapSecure`. Instead of
using the `Open()` method to set these, they are now directly set
using the corresponding `SecureSession` methods. This commit also
harmonizes the method name for setting the "connect" callback,
ensuring the same name is used by different classes.