[cli] allow executing the `factoryreset` command when running sync commands
When running the sync command, such as `diag radio receive 10000`, the cli
doesn't allow running other commands before the sync command returns. Which
causes that the sync command may be blocked for a very long time and users
can't do anything to terminate the sync command. This will also cause the
test script to fail to run the second case after running first case fails.

This commit allows the cli to execute the `factoryreset` command when running
the sync command.