[link-quality] track "link quality out" in `LinkQualityInfo`
This commit updates `LinkQualityInfo` to track the "link quality out,"
which is the link quality from the neighbor's perspective. This value
is currently tracked by the `Router` class. The existing methods in
`LinkQualityInfo` that return the "link quality in" are renamed to
use `LinkQualityIn` to clarify and distinguish them.

This change uses existing bits in `LinkQualityInfo`, freeing up bit
fields in the `Router` class to be used for other information.