[spinel] save the radio spinel metrics to temporary settings (#11058)
The otRadioSpinelMetrics includes metrics that record the count of the
RCP timeout and the count of RCP reboot and so on. After these events
happen, the Thread stack will exits and the system will re-start the
Thread stack again. In this case, metrics of the otRadioSpinelMetrics
will be cleared and metrics values are awalys 0.

This commit adds a temporary settings to store the radio spinel metrics.
The temporary settings will clear all settings after the system reboots.