[border-agent] organize `CoapDtlsSession` method definitions (#11142)
This commit improves the organization of `border_agent.cpp` by
grouping the method definitions for `CoapDtlsSession` into a separate
section.

In PR #11131, many methods were moved into this class. To keep the
`git diff` smaller and easier to review, the methods were left in the
same order as previously defined. This resulted in the methods of
`CoapDtlsSession` and `BorderAgent` being mixed and interleaved.

This PR rearranges the methods and defines separate sections for each
class. This commit does not contain any code logic changes.