[diag] correct usage of "diag frame"
The example for the "diag frame" command uses an invalid
802.15.4 frame while not bypassing the header and security
processing with "-s" flag.
In such a case, it cannot be guaranteed that the underlying
platform will send such a frame if it can't parse the frame
IEs to check if any dynamic data is to be injected.

Switch to using a correct ACK frame.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>